Add a meerkat-configs package for essentials hardware setup
The COM temperature sensor comes up with a default config that doesn't
match the thermal settings of the kernel. This package add a udev rule
that take care of configuring the temperature sensor according to the
kernel thermal settings. The temperature sensor is configured with a
slightly higher temperature limit than the kernel to have it act as
an emergency shutdown in case the software shutdown fails or isn't
fast enough.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>