kernel/linux: Disable WLAN support in the defconfig
We are going to use the backports drivers for WLAN, which in itself
could be used in parallel with drivers included in the kernel as long
as they are built as module. However the OE kernel receipes currently
don't properly support this and we end up with modules packages having
the same name. To solve this we just disable WLAN support in the
kernel, leaving these name free.

We also need to enable a few features in the kernel that are needed by
the backported drivers: LED support, the GCM crypto algorithm and the
averaging function.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>